build: migrate to autoconf, for the love of $DEITY
The GNU autotools exists in order to be actually portable and not invent
fake homemade build systems. Distribution packagers love it, because now
your ebuild can actually rely on the default src_configure! Also, no
need to pass manky variables constantly to other overridden phases.

This also means your Makefile can actually be for real portable, which
relying on stuff like `$(shell ... )` or for that matter even `:=` and
`?=` is simply... not.

The GNU autotools is actually really simple, too! We can do easy mode.
All we want is a configure script to make decisions for us. We can do
that in a couple of lines, and it runs blazingly fast too since nearly
all decisions are just parsing the command line into directories.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>